PM-74 move desktop hint component
